backend/requirements.txt
django==4.0.3
flake8==4.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0